Refactor helpers

+++ app/helpers/order_steps_helper.rb
@@ -1,5 +1,4 @@
 module OrderStepsHelper
-
   def last_order
     current_user.orders.in_queue.last
   end
@@ -7,10 +6,11 @@ module OrderStepsHelper
   private
 
   def validate_step
-    if session[previous_step.to_sym].nil? && step != steps.first
-      jump_to(previous_step.to_sym)
-    elsif step.eql? steps.last
-      steps.each { |s| session.delete(s) }
-    end
+    jump_to(previous_step.to_sym) if wrong_step?
+    steps.each { |s| session.delete(s) } if step.eql?(steps.last)
+  end
+
+  def wrong_step?
+    session[previous_step.to_sym].nil? && step != steps.first
   end
 end
